treewide: use root directories for all problems
Instead of splitting everything into a 'code' and 'data' folder at
the root, instead create folders at root for each problem. We will
populate each with a 'data' folder for the big data sets, but docs
and code will live side-by-side in those folders.

So far, we have the following problems that generate data:

* sz/sz.py : Szemeredi numbers
* vdw/vdw-2.py : 2-color van der Waerden numbers
* vdw/vdw-t.py : t-color van der Waerden numbers
* disc/disc.py : Erdos discrepancy problem

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>