sz: add data for Szemeredi numbers
The Szemeredi number sz(n,k) is the maximum size of a subset of
{1,...,n} that avoids k-arithmetic progressions. The Python script
sz/sz.py will compute these numbers (with a maximum k of 16) and
populate the sz/data directory with the Z3 scripts used, along with
a table of values in sz/data/sz.csv. The file is space-delimited.

The script will assume the sz/data/sz.csv file is correct and not
repeat values, so it will continue where it left off.

By running sz/sz.py and piping the output to sz/data/sz.out, we can
see the extremal examples.

The Z3 scripts can be run directly. They demonstrate the maximality
by providing an unsat proof when adding a constraint that the size
be strictly larger than the value sz(n,k).

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>