vdw: add data folder
van der Waerden numbers grow quickly, so our computations are geared
toward running one set of parameters at a time. The data folder
contains Z3 scripts for van der Waerden numbers with naming scheme

	vdw-[k values]-NNN.z3

For each tuple of k-values, there are two scripts with N-values off
by one. The lower number is for demonstrating existence of a good
coloring. The higher number is for demonstrating non-existence and
providing a proof (if desired).

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>