disc: add data folder for Erdos Discrepancy Problem
The extremal examples for the Erdos Discrepancy Problem grow very
quickly. In fact, for discrepancy C = 3, the value of 1125 was
only discovered recently (2015). Z3 can produce a proof of
extremality, but does take some time.

The data directory contains the Z3 scripts for finding extremal
examples and proofs of extremality. The output for the 'sat' and
'unsat' cases are provided for discrepancy C = 2.

A later commit will include the output for C = 3.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>