vdw: add a simple backtrack search
The new 'vdw' executable (written in C as vdw.c and compiled by
'make') performs a rudimentary backtrack search to compute van der
Waerden numbers. It also measures performance in terms of seconds
to compute the solution.

There are obvious inefficiencies in this code, but they can be
remedied in a later commit.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>