vdw: advanced propagation
In the backtrack search algorithm of vdw.c, use a more advanced
form of propagation using domains. These domains track the colors
available at a position.

If a domain becomes empty, then we can backtrack since the coloring
is impossible at this point.

If a domain contains exactly one value, then we can assign that
value to that position.

When we assign a color to a position, that can trigger another
propagation. Recursively propagate as necessary.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>